Refactor Global Float-to-Int Migration for Zero-Sum Integrity
This change implements a comprehensive migration from floating-point arithmetic to integer pennies across key financial protocols, DTOs, and handlers to ensure zero-sum integrity and eliminate "Financial Dust".

Key changes include:
- Protocol Update: `ICurrencyHolder` now mandates `get_balance() -> int` and `get_assets_by_currency() -> Dict[CurrencyCode, int]`.
- DTO Modernization: `GoodsInfoDTO`, `GoodsDTO`, `MarketContextDTO`, `MarketSignalDTO`, and `MarketHistoryDTO` now use `int` for all price-related fields.
- Logic Hardening: `InventoryLiquidationHandler` and `LiquidationManager` now enforce strict integer arithmetic for settlement transfers, removing legacy double-conversion logic.
- Bug Fix: Corrected `ProductionStrategy` to avoid redundant 100x price scaling in procurement orders.
- Default Value Standardization: Updated hardcoded fallback prices in `core_agents.py`, `firms.py`, and `utils.py` to integer pennies (e.g., 1000 instead of 10.0).
- Test Updates: Modernized `test_liquidation_waterfall.py` and unit tests to verify strict integer compliance and truncation behavior.

This resolves technical debt related to floating-point drift and enforces the "Penny Standard" across the simulation engine.

+++ modules/system/api.py
@@ -16,7 +16,7 @@ class MarketContextDTO:
     Contains strictly external market data (prices, rates, signals).
     """
     market_data: Dict[str, Any]
-    market_signals: Dict[str, float]
+    market_signals: Dict[str, int]
     tick: int
     # Represents currency exchange rates relative to base currency
     exchange_rates: Optional[Dict[str, float]] = None
@@ -25,11 +25,11 @@ class MarketContextDTO:
 class MarketSignalDTO:
     market_id: str
     item_id: str
-    best_bid: Optional[float]
-    best_ask: Optional[float]
-    last_traded_price: Optional[float]
+    best_bid: Optional[int]
+    best_ask: Optional[int]
+    last_traded_price: Optional[int]
     last_trade_tick: int
-    price_history_7d: List[float]
+    price_history_7d: List[int]
     volatility_7d: float
     order_book_depth_buy: int
     order_book_depth_sell: int
@@ -40,9 +40,9 @@ class MarketSignalDTO:
 @dataclass(frozen=True)
 class HousingMarketUnitDTO:
     unit_id: str
-    price: float
+    price: int
     quality: float
-    rent_price: Optional[float] = None
+    rent_price: Optional[int] = None
 
 @dataclass(frozen=True)
 class HousingMarketSnapshotDTO:
@@ -220,10 +220,10 @@ class ICurrencyHolder(Protocol):
     Protocol for agents/systems that hold currency.
     Used for M2 Money Supply calculation.
     """
-    def get_balance(self, currency: CurrencyCode = DEFAULT_CURRENCY) -> float:
+    def get_balance(self, currency: CurrencyCode = DEFAULT_CURRENCY) -> int:
         ...
 
-    def get_assets_by_currency(self) -> Dict[CurrencyCode, float]:
+    def get_assets_by_currency(self) -> Dict[CurrencyCode, int]:
         ...
 
 class IAssetRecoverySystem(Protocol):
